Initial implementation of rule aliases/named rules
Refs #99

Missing: Tests, Docs

Rules can be referred to by their aliases in --rules and in styles.
Add the -a option (or show_aliases in the config) to display rule names
instead of IDs in output.

This code supports multiple aliases for a rule, but currently only one alias
is present per rule. In addition, rules can have 0 aliases and their ID will
show instead.